tests(conftest): dispose of engines in the decorator to drop tables in tests (see notes)
We are creating lots of dormant connections in the database (it's always postgres that will complain) because one engine is being created per test and engines are not disposed of automatically.

I tried using `engine.dispose` in `pytest_generate_tests` but it did not work so I put that here.

There are still a lot of dormant connections but a lot less than the limit so this will do for now until we find a better fix that does not involve writing a decorator everywhere or such...